Organize components into /birds/ and /days/
