+++ src/App.js
@@ -3,8 +3,8 @@ import React from 'react'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import Nav from './components/Nav'
 import Home from './components/Home'
-import BirdsContainer from './components/BirdsContainer'
-import DaysContainer from './components/DaysContainer'
+import BirdsContainer from './components/birds/BirdsContainer'
+import DaysContainer from './components/days/DaysContainer'
 import { connect } from 'react-redux'
 import { fetchBirds } from './redux/actions/birdActions'
 import { fetchDays } from './redux/actions/dayActions'
